These are the Final Updates
So the final things I have done I had set up the game so that its ready to try, the scene is all set. The swinging didnt work so I made it go back to the ziping since it didnt bug oir have any issues and would let you move around the scene fluently. I also made it so that the player is high up and can mess around with the web shooter sticking on to wheverever they want too.